Example commit of mistral-server with implementing server-core
