+++ Cargo.toml
@@ -2,7 +2,6 @@
 members = [
     "mistralrs-server-core",
     "mistralrs-server",
-    "mistralrs-server-2",
     "mistralrs-core",
     "mistralrs-pyo3",
     "mistralrs",
