--- mistralrs-server-2/Cargo.toml
@@ -1,47 +1,0 @@
-[package]
-name = "mistralrs-server-2"
-readme = "README.md"
-authors = ["Eric Buehler"]
-version.workspace = true
-edition.workspace = true
-description.workspace = true
-repository.workspace = true
-keywords.workspace = true
-categories.workspace = true
-license.workspace = true
-homepage.workspace = true
-default-run = "mistralrs-server-2"
-
-[dependencies]
-anyhow.workspace = true
-axum = { workspace = true, features = ["tokio"] }
-clap.workspace = true
-ctrlc = "3.4.4"
-directories = "6.0.0"
-either.workspace = true
-indexmap.workspace = true
-mistralrs-core = { version = "0.6.0", path = "../mistralrs-core" }
-mistralrs-server-core = { version = "0.6.0", path = "../mistralrs-server-core" }
-once_cell.workspace = true
-regex.workspace = true
-rustyline = { version = "15.0.0", default-features = false, features = [
-  "with-file-history",
-] }
-serde.workspace = true
-serde_json.workspace = true
-tokio.workspace = true
-tracing.workspace = true
-
-[features]
-cuda = ["mistralrs-core/cuda", "mistralrs-server-core/cuda"]
-cudnn = ["mistralrs-core/cudnn", "mistralrs-server-core/cudnn"]
-metal = ["mistralrs-core/metal", "mistralrs-server-core/metal"]
-flash-attn = [
-  "cuda",
-  "mistralrs-core/flash-attn",
-  "mistralrs-server-core/flash-attn",
-]
-accelerate = ["mistralrs-core/accelerate", "mistralrs-server-core/accelerate"]
-mkl = ["mistralrs-core/mkl", "mistralrs-server-core/mkl"]
-nccl = ["mistralrs-core/nccl", "mistralrs-server-core/nccl"]
-ring = ["mistralrs-core/ring", "mistralrs-server-core/ring"]
