+++ mistralrs-server/src/main.rs
@@ -1,58 +1,15 @@
 use anyhow::Result;
-use axum::{
-    extract::{DefaultBodyLimit, Json, State},
-    http::{self, Method},
-    routing::{get, post},
-    Router,
-};
-use candle_core::Device;
 use clap::Parser;
-use mistralrs_core::{
-    get_auto_device_map_params, get_model_dtype, get_tgt_non_granular_index, initialize_logging,
-    paged_attn_supported, parse_isq_value, ApproximateUserLocation, BertEmbeddingModel,
-    DefaultSchedulerMethod, DeviceLayerMapMetadata, DeviceMapMetadata, DeviceMapSetting, Function,
-    ImageGenerationResponseFormat, Loader, LoaderBuilder, MemoryGpuConfig, MistralRs,
-    MistralRsBuilder, ModelSelected, PagedAttentionConfig, Request, SchedulerConfig,
-    SearchContextSize, TokenSource, Tool, ToolChoice, ToolType, WebSearchOptions,
-    WebSearchUserLocation,
-};
-use openai::{
-    AudioResponseFormat, ChatCompletionRequest, CompletionRequest, FunctionCalled, Grammar,
-    ImageGenerationRequest, JsonSchemaResponseFormat, Message, MessageContent, MessageInnerContent,
-    ModelObjects, ResponseFormat, SpeechGenerationRequest, StopTokens, ToolCall,
-};
-use serde::{Deserialize, Serialize};
-use speech_generation::{__path_speech_generation, speech_generation};
-use std::{num::NonZeroUsize, sync::Arc};
-
-mod chat_completion;
-mod completions;
-mod image_generation;
-mod interactive_mode;
-mod openai;
-mod speech_generation;
-mod util;
-
-use crate::openai::ModelObject;
-use crate::{
-    chat_completion::{__path_chatcompletions, chatcompletions},
-    completions::{__path_completions, completions},
-    image_generation::{__path_image_generation, image_generation},
-};
-
-use interactive_mode::interactive_mode;
-use tower_http::cors::{AllowOrigin, CorsLayer};
+use mistralrs_core::{ModelSelected, TokenSource};
 use tracing::info;
-use utoipa::{OpenApi, ToSchema};
-use utoipa_swagger_ui::SwaggerUi;
 
-// NOTE(EricLBuehler): Accept up to 50mb input
-const N_INPUT_SIZE: usize = 50;
-const MB_TO_B: usize = 1024 * 1024; // 1024 kb in a mb
+use mistralrs_server_core::{
+    mistralrs_for_server_builder::{defaults, get_bert_model, MistralRsForServerBuilder},
+    mistralrs_server_router_builder::MistralRsServerRouterBuilder,
+};
 
-fn parse_token_source(s: &str) -> Result<TokenSource, String> {
-    s.parse()
-}
+mod interactive_mode;
+use interactive_mode::interactive_mode;
 
 #[derive(Parser)]
 #[command(version, about, long_about = None)]
@@ -84,11 +41,11 @@ struct Args {
     model: ModelSelected,
 
     /// Maximum running sequences at any time. If the `tgt_non_granular_index` flag is set for X-LoRA models, this will be set to 1.
-    #[arg(long, default_value_t = 16)]
+    #[arg(long, default_value_t = defaults::MAX_SEQS)]
     max_seqs: usize,
 
     /// Use no KV cache.
-    #[arg(long, default_value_t = false)]
+    #[arg(long, default_value_t = defaults::NO_KV_CACHE)]
     no_kv_cache: bool,
 
     /// Chat template file with a JINJA file with `messages`, `add_generation_prompt`, `bos_token`, `eos_token`, and `unk_token` as inputs.
@@ -103,7 +60,7 @@ struct Args {
     /// Source of the token for authentication.
     /// Can be in the formats: `literal:<value>`, `env:<value>`, `path:<value>`, `cache` to use a cached token, or `none` to use no token.
     /// Defaults to `cache`.
-    #[arg(long, default_value_t = TokenSource::CacheToken, value_parser = parse_token_source)]
+    #[arg(long, default_value_t = defaults::TOKEN_SOURCE, value_parser = parse_token_source)]
     token_source: TokenSource,
 
     /// Enter interactive mode instead of serving a chat server.
@@ -111,7 +68,7 @@ struct Args {
     interactive_mode: bool,
 
     /// Number of prefix caches to hold on the device. Other caches are evicted to the CPU based on a LRU strategy.
-    #[arg(long, default_value_t = 16)]
+    #[arg(long, default_value_t = defaults::PREFIX_CACHE_N)]
     prefix_cache_n: usize,
 
     /// NOTE: This can be omitted to use automatic device mapping!
@@ -151,11 +108,11 @@ struct Args {
     paged_attn_block_size: Option<usize>,
 
     /// Disable PagedAttention on CUDA. Because PagedAttention is already disabled on Metal, this is only applicable on CUDA.
-    #[arg(long = "no-paged-attn", default_value_t = false)]
+    #[arg(long = "no-paged-attn", default_value_t = defaults::NO_PAGED_ATTN)]
     no_paged_attn: bool,
 
     /// Enable PagedAttention on Metal. Because PagedAttention is already enabled on CUDA, this is only applicable on Metal.
-    #[arg(long = "paged-attn", default_value_t = false)]
+    #[arg(long = "paged-attn", default_value_t = defaults::PAGED_ATTN)]
     paged_attn: bool,
 
     /// Number of tokens to batch the prompt step into. This can help with OOM errors when in the prompt step, but reduces performance.
@@ -179,342 +136,42 @@ struct Args {
     enable_thinking: bool,
 }
 
-#[utoipa::path(
-    get,
-    tag = "Mistral.rs",
-    path = "/v1/models",
-    responses((status = 200, description = "Served model info", body = ModelObjects))
-)]
-async fn models(State(state): State<Arc<MistralRs>>) -> Json<ModelObjects> {
-    Json(ModelObjects {
-        object: "list",
-        data: vec![ModelObject {
-            id: state.get_id(),
-            object: "model",
-            created: state.get_creation_time(),
-            owned_by: "local",
-        }],
-    })
-}
-
-#[utoipa::path(
-    get,
-    tag = "Mistral.rs",
-    path = "/health",
-    responses((status = 200, description = "Server is healthy"))
-)]
-async fn health() -> &'static str {
-    "OK"
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema)]
-struct ReIsqRequest {
-    #[schema(example = "Q4K")]
-    ggml_type: String,
-}
-
-#[utoipa::path(
-    post,
-    tag = "Mistral.rs",
-    path = "/re_isq",
-    request_body = ReIsqRequest,
-    responses((status = 200, description = "Reapply ISQ to a non GGUF or GGML model."))
-)]
-async fn re_isq(
-    State(state): State<Arc<MistralRs>>,
-    Json(request): Json<ReIsqRequest>,
-) -> Result<String, String> {
-    let repr = format!("Re ISQ: {:?}", request.ggml_type);
-    MistralRs::maybe_log_request(state.clone(), repr.clone());
-    let request = Request::ReIsq(parse_isq_value(&request.ggml_type, None)?);
-    state.get_sender().unwrap().send(request).await.unwrap();
-    Ok(repr)
-}
-
-fn get_router(state: Arc<MistralRs>) -> Router {
-    #[derive(OpenApi)]
-    #[openapi(
-        paths(models, health, chatcompletions, completions, re_isq, image_generation, speech_generation),
-        components(schemas(
-            ApproximateUserLocation,
-            AudioResponseFormat,
-            ChatCompletionRequest,
-            CompletionRequest,
-            Function,
-            FunctionCalled,
-            Grammar,
-            ImageGenerationRequest,
-            ImageGenerationResponseFormat,
-            JsonSchemaResponseFormat,
-            Message,
-            MessageContent,
-            MessageInnerContent,
-            ModelObject,
-            ModelObjects,
-            ReIsqRequest,
-            ResponseFormat,
-            SearchContextSize,
-            SpeechGenerationRequest,
-            StopTokens,
-            Tool,
-            ToolCall,
-            ToolChoice,
-            ToolType,
-            WebSearchOptions,
-            WebSearchUserLocation
-        )),
-        tags(
-            (name = "Mistral.rs", description = "Mistral.rs API")
-        ),
-        info(
-            title = "Mistral.rs",
-            license(
-            name = "MIT",
-        )
-        )
-    )]
-    struct ApiDoc;
-
-    let doc = { ApiDoc::openapi() };
-
-    let allow_origin = AllowOrigin::any();
-    let cors_layer = CorsLayer::new()
-        .allow_methods([Method::GET, Method::POST])
-        .allow_headers([http::header::CONTENT_TYPE, http::header::AUTHORIZATION])
-        .allow_origin(allow_origin);
-
-    Router::new()
-        .merge(SwaggerUi::new("/docs").url("/api-doc/openapi.json", doc))
-        .route("/v1/chat/completions", post(chatcompletions))
-        .route("/v1/completions", post(completions))
-        .route("/v1/models", get(models))
-        .route("/health", get(health))
-        .route("/", get(health))
-        .route("/re_isq", post(re_isq))
-        .route("/v1/images/generations", post(image_generation))
-        .route("/v1/audio/speech", post(speech_generation))
-        .layer(cors_layer)
-        .layer(DefaultBodyLimit::max(N_INPUT_SIZE * MB_TO_B))
-        .with_state(state)
+fn parse_token_source(s: &str) -> Result<TokenSource, String> {
+    s.parse()
 }
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    let mut args = Args::parse();
-    initialize_logging();
+    let args = Args::parse();
 
-    let tgt_non_granular_index = get_tgt_non_granular_index(&args.model);
-    let dtype = get_model_dtype(&args.model)?;
-    let auto_device_map_params = get_auto_device_map_params(&args.model)?;
-
-    if tgt_non_granular_index.is_some() {
-        args.max_seqs = 1;
-    }
-
-    let prompt_chunksize = match args.prompt_chunksize {
-        Some(0) => {
-            anyhow::bail!("`prompt_chunksize` must be a strictly positive integer, got 0.",)
-        }
-        Some(x) => Some(NonZeroUsize::new(x).unwrap()),
-        None => None,
-    };
-
-    let max_seq_len = auto_device_map_params.max_seq_len();
-
-    let loader: Box<dyn Loader> = LoaderBuilder::new(args.model)
+    let mistralrs = MistralRsForServerBuilder::new()
+        .with_truncate_sequence(args.truncate_sequence)
+        .with_model(args.model)
+        .with_max_seqs(args.max_seqs)
         .with_no_kv_cache(args.no_kv_cache)
-        .with_chat_template(args.chat_template)
-        .with_prompt_chunksize(prompt_chunksize)
-        .with_jinja_explicit(args.jinja_explicit)
-        .build()?;
+        .with_token_source(args.token_source)
+        .with_interactive_mode(args.interactive_mode)
+        .with_prefix_cache_n(args.prefix_cache_n)
+        .with_no_paged_attn(args.no_paged_attn)
+        .with_paged_attn(args.paged_attn)
+        .with_cpu(args.cpu)
+        .with_enable_search(args.enable_search)
+        .with_seed_optional(args.seed)
+        .with_log_optional(args.log)
+        .with_chat_template_optional(args.chat_template)
+        .with_jinja_explicit_optional(args.jinja_explicit)
+        .with_num_device_layers_optional(args.num_device_layers)
+        .with_in_situ_quant_optional(args.in_situ_quant)
+        .with_paged_attn_gpu_mem_optional(args.paged_attn_gpu_mem)
+        .with_paged_attn_gpu_mem_usage_optional(args.paged_attn_gpu_mem_usage)
+        .with_paged_ctxt_len_optional(args.paged_ctxt_len)
+        .with_paged_attn_block_size_optional(args.paged_attn_block_size)
+        .with_prompt_chunksize_optional(args.prompt_chunksize)
+        .build()
+        .await?;
 
-    #[cfg(feature = "metal")]
-    let device = if args.cpu {
-        args.no_paged_attn = true;
-        Device::Cpu
-    } else {
-        Device::new_metal(0)?
-    };
-    #[cfg(not(feature = "metal"))]
-    let device = if args.cpu {
-        args.no_paged_attn = true;
-        Device::Cpu
-    } else if mistralrs_core::distributed::use_nccl() {
-        Device::Cpu
-    } else {
-        Device::cuda_if_available(0)?
-    };
-
-    if let Some(seed) = args.seed {
-        device.set_seed(seed)?;
-    }
-
-    info!(
-        "avx: {}, neon: {}, simd128: {}, f16c: {}",
-        candle_core::utils::with_avx(),
-        candle_core::utils::with_neon(),
-        candle_core::utils::with_simd128(),
-        candle_core::utils::with_f16c()
-    );
-    info!("Sampling method: penalties -> temperature -> topk -> topp -> minp -> multinomial");
-    info!("Model kind is: {}", loader.get_kind().to_string());
-
-    // Parse device mapper
-    let mapper = if let Some(device_layers) = args.num_device_layers {
-        if device_layers.len() == 1 && device_layers[0].parse::<usize>().is_ok() {
-            let layers = device_layers[0].parse::<usize>().unwrap();
-            DeviceMapSetting::Map(DeviceMapMetadata::from_num_device_layers(vec![
-                DeviceLayerMapMetadata { ordinal: 0, layers },
-            ]))
-        } else {
-            let mut mapping = Vec::new();
-            for layer in device_layers {
-                let split = layer.splitn(2, ':').collect::<Vec<_>>();
-                if split.len() < 2 {
-                    panic!("Expected layer to be of format ORD:NUM, got {layer}");
-                }
-                let ord = split[0]
-                    .parse::<usize>()
-                    .unwrap_or_else(|_| panic!("Failed to parse {} as integer.", split[0]));
-                let num = split[1]
-                    .parse::<usize>()
-                    .unwrap_or_else(|_| panic!("Failed to parse {} as integer.", split[1]));
-                for DeviceLayerMapMetadata { ordinal, layers: _ } in &mapping {
-                    if *ordinal == ord {
-                        panic!("Duplicate ordinal {ord}");
-                    }
-                }
-                mapping.push(DeviceLayerMapMetadata {
-                    ordinal: ord,
-                    layers: num,
-                });
-            }
-            DeviceMapSetting::Map(DeviceMapMetadata::from_num_device_layers(mapping))
-        }
-    } else {
-        DeviceMapSetting::Auto(auto_device_map_params)
-    };
-
-    let no_paged_attn = if device.is_cuda() || mistralrs_core::distributed::use_nccl() {
-        args.no_paged_attn
-    } else if device.is_metal() {
-        !args.paged_attn
-    } else {
-        true
-    };
-
-    // Allocate 0.5 GB of CPU memory just as a placeholder.
-    // Nothing happens here as we have no `swap_out`, see `_preempt_by_swap`.
-    let cache_config = match (
-        args.paged_attn_block_size,
-        args.paged_attn_gpu_mem,
-        args.paged_attn_gpu_mem_usage,
-        args.paged_ctxt_len,
-        paged_attn_supported(),
-        no_paged_attn,
-    ) {
-        (block_size, None, None, None, true, false) => Some(PagedAttentionConfig::new(
-            block_size,
-            512,
-            MemoryGpuConfig::ContextSize(max_seq_len),
-        )?),
-        (block_size, None, None, Some(ctxt), true, false) => Some(PagedAttentionConfig::new(
-            block_size,
-            512,
-            MemoryGpuConfig::ContextSize(ctxt),
-        )?),
-        (block_size, None, Some(f), None, true, false) => Some(PagedAttentionConfig::new(
-            block_size,
-            512,
-            MemoryGpuConfig::Utilization(f),
-        )?),
-        (block_size, Some(m), None, None, true, false) => Some(PagedAttentionConfig::new(
-            block_size,
-            512,
-            MemoryGpuConfig::MbAmount(m),
-        )?),
-        (block_size, Some(_m), Some(f), None, true, false) => {
-            info!("Both memory size, and usage were specified, defaulting to the usage value.");
-            Some(PagedAttentionConfig::new(
-                block_size,
-                512,
-                MemoryGpuConfig::Utilization(f),
-            )?)
-        }
-        (block_size, Some(_m), None, Some(ctxt), true, false) => {
-            info!("All memory size and ctxt len, defaulting to the context len value.");
-            Some(PagedAttentionConfig::new(
-                block_size,
-                512,
-                MemoryGpuConfig::ContextSize(ctxt),
-            )?)
-        }
-        (block_size, None, Some(f), Some(_ctxt), true, false) => {
-            info!("Both ctxt len and usage were specified, defaulting to the usage value.");
-            Some(PagedAttentionConfig::new(
-                block_size,
-                512,
-                MemoryGpuConfig::Utilization(f),
-            )?)
-        }
-        (_, _, _, _, _, _) => None,
-    };
-
-    let isq = args
-        .in_situ_quant
-        .as_ref()
-        .and_then(|isq| parse_isq_value(isq, Some(&device)).ok());
-
-    let pipeline = loader.load_model_from_hf(
-        None,
-        args.token_source,
-        &dtype,
-        &device,
-        false,
-        mapper,
-        isq,
-        cache_config,
-    )?;
-    info!("Model loaded.");
-
-    let scheduler_config = if cache_config.is_some() {
-        // Handle case where we may have device mapping
-        if let Some(ref cache_config) = pipeline.lock().await.get_metadata().cache_config {
-            SchedulerConfig::PagedAttentionMeta {
-                max_num_seqs: args.max_seqs,
-                config: cache_config.clone(),
-            }
-        } else {
-            SchedulerConfig::DefaultScheduler {
-                method: DefaultSchedulerMethod::Fixed(args.max_seqs.try_into().unwrap()),
-            }
-        }
-    } else {
-        SchedulerConfig::DefaultScheduler {
-            method: DefaultSchedulerMethod::Fixed(args.max_seqs.try_into().unwrap()),
-        }
-    };
-    let bert_model = if args.enable_search {
-        Some(
-            args.search_bert_model
-                .map(BertEmbeddingModel::Custom)
-                .unwrap_or_default(),
-        )
-    } else {
-        None
-    };
-    // Throughput logging in the server
-    let mistralrs = MistralRsBuilder::new(
-        pipeline,
-        scheduler_config,
-        !args.interactive_mode,
-        bert_model.clone(),
-    )
-    .with_opt_log(args.log)
-    .with_truncate_sequence(args.truncate_sequence)
-    .with_no_kv_cache(args.no_kv_cache)
-    .with_prefix_cache_n(args.prefix_cache_n)
-    .build();
+    // TODO: refactor this
+    let bert_model = get_bert_model(args.enable_search, args.search_bert_model);
 
     if args.interactive_mode {
         interactive_mode(
@@ -538,7 +195,11 @@ async fn main() -> Result<()> {
         None
     };
 
-    let app = get_router(mistralrs);
+    let app = MistralRsServerRouterBuilder::new()
+        .with_mistralrs(mistralrs)
+        .build()
+        .await?;
+
     if let Some((listener, ip, port)) = setting_server {
         info!("Serving on http://{ip}:{}.", port);
         axum::serve(listener, app).await?;
